Remove pool upgrade option
